Add port in Meta to inherit remote server port from discovery

+++ pkg/mod/models.go
@@ -21,7 +21,7 @@ type (
 
 	ServerConfig struct {
 		Bind string `mapstructure:"bind,omitempty"`
-		Port int    `mapstructure:"port,omitempty"`
+		Port uint16    `mapstructure:"port,omitempty"`
 	}
 
 	DiscoveryConfig struct {
